Make PDP-11 API use 32-bit words instead of 16-bit words, packing 6 characters into each word.
This is a breaking change.

The PDP-11 API is still 16 bit which causes some friction in the places
where they want to share code. There's no built-in Number trait in std, so
simple generic functions won't cut it. Instead there's now a GenerticCodec
trait that both the pdp10 and pdp11 modules have impls of.

The cli takes a different tack. When encoding, it converts both types of
encode into a u64 array and then in the raw format case hardcodes the 16/32
bit conversion. When decoding it has a generic function that has a const
parameter for number of bytes per word. It _should_ have been able to
auto-detect this but apparently there is a limitation of rustc that prevents
it.

Closes 2