cli: Change exe name back to radix50 to match docs
Looks like it changed back in 8163f4b76092929ce8a98e153cefcbb831ee915a and
not mentioned in the commit message which makes me think the commit was not
intentional. I have no memory of changing it :-). I think huffman encoding
radix50 down to r50 is a bit of a premature optimization for something that
is a niche tool at best.

Sorry if anyone had previously actually `cargo install`ed it.